+++ client/src/components/NavBar.css
@@ -7,6 +7,7 @@
   fill: var(--clr-text);
   height: 1.5rem;
   width: auto;
+  display: block;
 }
 
 .theme-toggle {
